third_party/blink/web_tests/fast/dom/Window/window-resize-contents.html

<!DOCTYPE html>
<html>
<body>
<div id="reference" style="position: absolute; top: 0; bottom: 0; left: 0; right: 0; background: yellow; z-index: 1"></div>

<div style="position: absolute; z-index: 2">
    <p>This test checks that the yellow reference DOM node (which should be as
    big as the window) gets resized when the window is resized.</p>
</div>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
promise_test(async () => {
  var referenceNode = document.getElementById('reference');
  var originalWidth = referenceNode.offsetWidth;
  var originalHeight = referenceNode.offsetHeight;
  var resizePromise = new Promise(resolve => window.onresize = resolve);
  window.resizeBy(10, 10);
  await resizePromise;

  assert_equals(originalWidth + 10, referenceNode.offsetWidth);
  assert_equals(originalHeight + 10, referenceNode.offsetHeight);
});
</script>

</body>
</html>
